backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "fastapi-pagination>=0.15.0",
    "fastapi-sso>=0.18.0",
    "pydantic>=2.12.5",
    "python-decouple>=3.8",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "ruff>=0.14.6",
]
